db_refresh_pecas_views/requirements.txt
requests
pandas
numpy
python-dotenv
click
sqlalchemy
psycopg2-binary
geopandas
shapely
geojson
protobuf
xlsxwriter
unidecode
